#71: Move a function to the main file
